Moved EitherOrBoth  to the Funcky root namespace * Remove the DataTypes namespace
