+++ Funcky/EitherOrBoth.cs
@@ -1,6 +1,6 @@
 using Funcky.Internal;
 
-namespace Funcky.DataTypes
+namespace Funcky
 {
     /// <remarks>
     /// EitherOrBoth values constructed using <c>default</c> are in an invalid state.
